pyproject.toml
[tool.poetry]
name = "signs"
version = "0.1.0"
description = "整理腾讯在线收集表的签名。"
authors = ["Y.D.X. <73375426+YDX-2147483647@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openpyxl = "^3.1.2"
openpyxl-image-loader = "^1.0.5"
numpy = "^1.25.2"
opencv-python = "^4.8.0.74"
fpdf2 = "^2.7.4"
pypdf = "^3.14.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # Pyflakes
    "I", # isort
    "N", # pep8-naming
]
